FoodieTrack/backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
pyjwt[crypto]>=2.8.0
httpx>=0.27.0
google-genai>=1.0.0
pydantic>=2.0.0
